EksOppg/requirements.txt
bidict==0.21.4
click==8.0.3
colorama==0.4.4
dominate==2.6.0
Flask==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==5.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
mysql-connector==2.2.9
pyFirmata==1.1.0
pyserial==3.5
python-engineio==4.3.1
python-socketio==5.5.1
visitor==0.1.3
Werkzeug==2.0.3
